Summary
This code uses the Giphy API to search for GIFs based on the user's input. When the user submits the form, the code makes an AJAX request to the Giphy API and returns a single GIF. The GIF is then appended to the #gif-container element. The user can continue to search for GIFs and they will be appended to the page. The user can also click the #clear-gifs button to remove all of the GIFs from the page.